Manage sources from extension details (closes #3152)
