--- app/src/main/res/layout/extension_preferences_controller.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/extension_prefs_recycler"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent" />
-
-    <eu.kanade.tachiyomi.widget.EmptyView
-        android:id="@+id/extension_prefs_empty_view"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:gravity="center"
-        android:visibility="gone"
-        tools:visibility="visible" />
-
-</FrameLayout>
